requirements.txt
asgiref==3.5.2
autopep8==2.0.0
beaker-pyteal==0.4.0
cffi==1.15.1
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
docstring-parser==0.14.1
Markdown==3.4.1
msgpack==1.0.4
mysqlclient==2.1.1
py-algorand-sdk==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.16.0
PyNaCl==1.5.0
pyteal==0.20.1
pytz==2022.6
semantic-version==2.10.0
sqlparse==0.4.3
tomli==2.0.1
xlwt==1.3.0
